Re-implement core to gain performance and simplify code #44 - separated Node and EvolvingNode
